organize ca_elements for lib
