+++ gui/mainwindow.py
@@ -1,7 +1,7 @@
 import sys
 import os
 import math
-from lib.main.main import CAFile
+from lib.ca_elements.core import CAFile
 from PySide6 import QtCore
 from PySide6.QtCore import Qt, QRectF, QPointF, QSize, QEvent, QVariantAnimation, QKeyCombination, QKeyCombination
 from PySide6.QtGui import QPixmap, QImage, QBrush, QPen, QColor, QTransform, QPainter, QLinearGradient, QIcon, QPalette, QFont, QShortcut, QKeySequence
