+++ lib/ca_elements/animation/__init__.py
@@ -1,0 +1,4 @@
+from .caanimation import CAAnimation
+from .caspringanimation import CASpringAnimation
+from .camatchmoveanimation import CAMatchMoveAnimation
+from .cakeyframeanimation import CAKeyframeAnimation
