+++ lib/ca_elements/core/cgimage.py
@@ -1,0 +1,3 @@
+class CGImage:
+    def __init__(self, src):
+        self.src = src
