+++ lib/ca_elements/core/cgpoint.py
@@ -1,0 +1,14 @@
+import xml.etree.ElementTree as ET
+
+class CGPoint:
+    def __init__(self, element):
+        self.element = element
+        self.type = self.element.tag.replace(
+            "{http://www.apple.com/CoreAnimation/1.0}", "")
+        self.value = self.element.get("value")
+
+    def create(self):
+        e = ET.Element(self.type)
+        e.set("value", self.value)
+
+        return e
